+++ __fpathperror2.c
@@ -22,10 +22,9 @@ void __fpathperror2(int fd1, int fd2, const char *s)
 	siz2 = fpath(fd2, buf2, sizeof(buf2));
 	if (siz1 == -1 || siz2 == -1) {
 #ifdef __FreeBSD__
-		__notice("%s: %i, %i: %s: %i\n", __getrootdir(), fd1, fd2, s,
-			 save_errno);
+		__notice("%i, %i: %s: %i\n", fd1, fd2, s, save_errno);
 #else
-		__notice("%s: %i, %i: %s: %s\n", __getrootdir(), fd1, fd2, s,
+		__notice("%i, %i: %s: %s\n", fd1, fd2, s,
 			 strerror(save_errno));
 #endif
 		errno = save_errno;
@@ -34,20 +33,20 @@ void __fpathperror2(int fd1, int fd2, const char *s)
 
 	if (__ignored_errno(errno) || __ignored_function(s)) {
 #ifdef __FreeBSD__
-		__debug("%s: %i <-> %s, %i <-> %s: %s: %i\n", __getrootdir(),
-			fd1, buf1, fd2, buf2, s, errno);
+		__debug("%i <-> %s, %i <-> %s: %s: %i\n", fd1, buf1, fd2, buf2,
+			s, errno);
 #else
-		__debug("%s: %i <-> %s, %i <-> %s: %s: %m\n", __getrootdir(),
-			fd1, buf1, fd2, buf2, s);
+		__debug("%i <-> %s, %i <-> %s: %s: %m\n", fd1, buf1, fd2, buf2,
+			s);
 #endif
 		return;
 	}
 
 #ifdef __FreeBSD__
-	__note_or_fatal("%s: %i <-> %s, %i <-> %s: %s: %i\n", __getrootdir(),
-			fd1, buf1, fd2, buf2, s, errno);
+	__note_or_fatal("%i <-> %s, %i <-> %s: %s: %i\n", fd1, buf1, fd2, buf2,
+			s, errno);
 #else
-	__note_or_fatal("%s: %i <-> %s, %i <-> %s: %s: %m\n", __getrootdir(),
-			fd1, buf1, fd2, buf2, s);
+	__note_or_fatal("%i <-> %s, %i <-> %s: %s: %m\n", fd1, buf1, fd2, buf2,
+			s);
 #endif
 }